requirements.txt
openai==0.28.1
python-dotenv==1.0.0
panel==1.3.8
ipython
jupyter
numpy
pandas